Upgrade template (#33)
### Rebuilt Sanity Nuxt Clean Template

This update completely rebuilds the [Sanity Nuxt Clean Template](https://github.com/sanity-io/sanity-template-nuxt-clean) from the ground up. It helpful defaults that make it easier to get started—or take things further.

#### ✨ What’s New

- **Modern frontend stack**
  - Built with [Nuxt 3](https://nuxt.com/)
  - Styled using [Tailwind CSS](https://tailwindcss.com/)
  - Dynamic pages powered by a simple page builder
  - Includes a basic blog as a starting point

- **Improved Sanity Studio**
  - Schemas for `Pages`, `Posts`, and basic `Settings`
  - [Presentation tool](https://www.sanity.io/docs/configuring-the-presentation-tool) set up for live visual editing
  - [TypeGen](https://www.sanity.io/docs/sanity-typegen) enabled for schema type safety
  - [AI Assist](https://www.sanity.io/ai-assist) basics included (e.g. image captioning, available during Growth plan trial)

#### 🎯 Why This Matters

This template gives you a strong foundation to build on. It’s opinionated enough to guide you, but simple enough to extend—or strip down—based on your needs.

#### 📦 Keeping the Old Version

If you need the previous version of the template, it’s now available on a new branch (or in the git history of main):  
[`pre-template-refactor`](https://github.com/sanity-io/sanity-template-nuxt-clean/tree/pre-template-refactor)

#### 🙌 Shoutout

Big thanks to ChrisLaRocque <larocque.christopher@gmail.com>, testing, and code reviews